Implement API endpoint to retrieve distribution of books by genre
Implemented a new API endpoint getGenreDistribution to retrieve the distribution of books by genre. The endpoint queries the database to count the number of books in each genre using Django's Count function. It then generates a bar chart to visualize the distribution and saves it as an image file. The file path of the generated plot is returned in the API response.